Nicer scrolling behaviour when removing lines from the bottom.
Got the nicer behaviour by adding padding to the bottom the size of the
whole document. This way if the user deletes text, the caret is allowed to
move up the screen, rather than have the scrollbar immediately shorten so
that the caret stays at the bottom of the screen.